NMR-6891: inline MainApp in AnalystApp, replace calls to MainApp
