+++ nmrfx-analyst-gui/src/main/java/org/nmrfx/analyst/gui/AnalystPrefs.java
@@ -13,7 +13,8 @@ import javafx.beans.property.IntegerProperty;
 import javafx.beans.property.StringProperty;
 import org.controlsfx.control.PropertySheet;
 import org.nmrfx.chemistry.io.PDBFile;
-import static org.nmrfx.processor.gui.MainApp.preferencesController;
+
+import static org.nmrfx.analyst.gui.AnalystApp.preferencesController;
 import org.nmrfx.processor.gui.PreferencesController;
 import org.nmrfx.utils.properties.BooleanOperationItem;
 import org.nmrfx.utils.properties.DirectoryOperationItem;
